requirements.txt
ollama
transformers
python-dotenv
sentence-transformers
faiss-cpu

--extra-index-url https://download.pytorch.org/whl/cu124
torch
torchvision
torchaudio